Update * Updated brewfile for current apps * Updated CreateBrewFile to be more verbose

+++ CreateBrewFile.sh
@@ -1,5 +1,11 @@
-# Remove current BrewFile
-rm BrewFile -v
+# Check if BrewFile exists and remove it if it does
+if [ -e "BrewFile" ]; then
+  echo "Removing existing BrewFile"
+  rm BrewFile -v
+else
+  echo "BrewFile does not exist"
+fi
 
 # Create BrewFile
-brew bundle dump
+echo "Creating BrewFile"
+brew bundle dump -v
